% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_timeline.r
\name{deployment_timeline}
\alias{deployment_timeline}
\title{Create a Timeline Showing any Receiver Failures}
\usage{
deployment_timeline(x, filename, path = NULL, create.dir = FALSE)
}
\arguments{
\item{x}{a dataframe created by the \code{\link[=read_sg_gps]{read_sg_gps()}} function.}

\item{filename}{File name to create on disk.}

\item{path}{Path of the directory to save plot to: path and filename are combined to create the fully qualified file name. Defaults to the working directory.}

\item{create.dir}{Whether to create new directories if a non-existing directory is specified in the filename or path (TRUE) or return an error (FALSE, default). If FALSE and run in an interactive session, a prompt will appear asking to create a new directory when necessary.}
}
\description{
This function creates a timeline of GPS hits (read from raw data using the \code{\link[=read_sg_gps]{read_sg_gps()}} function) that can be used to determine any gaps in data collection resulting from power issues or other problems. Normally, GPS hits are collected by the Sensorgnome every minute, so significant gaps can indicate equipment issues. This function is designed to recreate part of the Deployment Timeline available for station managers on the Motus website. The timeline is exported as an image.
}
