% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sg.R
\name{extract_sg}
\alias{extract_sg}
\title{Extract and Compile Zipped Sensorgnome Data}
\usage{
extract_sg(folder_path, output_file)
}
\arguments{
\item{folder_path}{the path to the extracted folder with the Sensorgnome data. It should look like: "SG-1AAARPI3A1AB-2026-07-01T14_20_55.123Z/".}

\item{output_file}{the desired name of the output file. Don't forget to include the ".txt" extension at the end.}
}
\description{
This functions compiles the zipped folders in a Sensorgnome data package and compiles them into a .txt file.
}
