% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sg.R
\name{read_sg}
\alias{read_sg}
\title{Read and Clean Compiled Files}
\usage{
read_sg(x, keep_only_ports = TRUE)
}
\arguments{
\item{x}{the file name and path of the .txt file you want to read.}

\item{keep_only_ports}{logical. Specifies if you want to filter out all data except pulse data. Defaults to TRUE. It is highly recommended that you do not change this parameter unless you have experience working with this data.}
}
\description{
This function reads and cleans the compiled .txt file produced by the \code{\link[=extract_sg]{extract_sg()}} function.
}
