% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runlen_filter.R
\name{runlen_filter}
\alias{runlen_filter}
\title{Create a Run Length Filter}
\usage{
runlen_filter(x, burst_interval, gap = 13)
}
\arguments{
\item{x}{a data frame created by the \code{\link[=search_code]{search_code()}} function.}

\item{burst_interval}{the tag's burst interval (eg. 6.7 or 15.1).}

\item{gap}{the number of bursts that can be missed before a new run starts. Defaults to 13.}
}
\description{
This function adds a run length filter comparable to the "motusFilter" to a dataframe produced by \code{\link[=search_code]{search_code()}}.
}
