% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_code.R
\name{search_code}
\alias{search_code}
\title{Search Sensorgnome Data for a Known Codeset}
\usage{
search_code(
  x,
  code,
  pulse_1,
  pulse_2,
  pulse_3,
  upper_freq = 5,
  lower_freq = 3,
  signal_duration = 0.015,
  S2N_cutoff = 8,
  tz = "UTC",
  create_csv = FALSE,
  output_file = NULL,
  output_dir = NULL
)
}
\arguments{
\item{x}{a data frame created by the \code{\link[=read_sg]{read_sg()}} function.}

\item{code}{the tag's 3-number code.}

\item{pulse_1}{the first pulse gap in seconds.}

\item{pulse_2}{the second pulse gap in seconds.}

\item{pulse_3}{the third pulse gap in seconds.}

\item{upper_freq}{the upper frequency offset threshold. Defaults to 5.}

\item{lower_freq}{the upper frequency offset threshold. Defaults to 3.}

\item{signal_duration}{the minimum signal duration cutoff. Defaults to 0.015 milliseconds.}

\item{S2N_cutoff}{the minimum signal to noise ratio allowed. Defaults to 8.}

\item{tz}{specifies which timezone you want the data to output in. A character string. The time zone specification to be used for the conversion, if one is required. System-specific (\code{\link{time zones}}), but "" is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning. Defaults to "UTC"}

\item{create_csv}{logical. If TRUE, output_file and output_dir must be specified. Defaults to FALSE.}

\item{output_file}{the name of the .csv file to be created.}

\item{output_dir}{the directory you want the file to be stored in. If no directory is specified, the file will be created in the working directory.}
}
\description{
This function searches through Sensorgnome data processed with the [read_sg())] to find the specific pulse code set associated with a tag.

[read_sg())]: R:read_sg())
}
\details{
This process is not meant to replace the process used by the broader Motus network, but is simply meant to be used to quickly process raw Sensorgnome data to diagnose deployment issues or search known tags quickly.
}
